Comments

+++ morse.cpp
@@ -1,14 +1,17 @@
+// Include headerfiles 
 #include "morse.h"
 #include "bst.h"
 #include "queue.h"
 
+// Include Libraries
 #include <ctype.h>
 #include <stdio.h>
 #include <string.h>
 
+// Allocate memory to stack
 Stack path = (Stack)malloc(sizeof(struct stack));
 
-void toNum(char sentance[], Queue *queue) {
+void toNum(char sentance[], Queue *queue) { // Function that seperates input into words and creates sendable array
     char word[ARR_SIZE] = {END};
     int word_int[ARR_SIZE] = {END};
 
@@ -17,10 +20,10 @@ void toNum(char sentance[], Queue *queue) {
         if (sentance[i] == NEWLINE || sentance[i] == 47) {
             sentance[i] = END;
         }
-        if (sentance[i] != SPACE && sentance[i] != END) {
+        if (sentance[i] != SPACE && sentance[i] != END) { // Adds letter to word array if it is not the end of a word or sentance
             word[j] = sentance[i];
 
-        } else {
+        } else { // Translates and adds word to the queue
             word[j] = WORDEND;
             translate(word, word_int);
             Enqueue(queue, CreateItem(word_int));
@@ -32,14 +35,14 @@ void toNum(char sentance[], Queue *queue) {
     }
 }
 
-void translate(char word[], int num[]) {
-    if (word[0] == DOT || word[0] == DASH) {
+void translate(char word[], int num[]) { // Translates one word in text or morse to decimal values
+    if (word[0] == DOT || word[0] == DASH) { // Checks if input is text or morse code
         int temp[4] = {END};
         int count = 0;
-        for (int i = 0, j = 0; i < ARR_SIZE && word[i] != END; i++, j++) {
-            if (word[i] != WORDEND) {
+        for (int i = 0, j = 0; i < ARR_SIZE && word[i] != END; i++, j++) { // sorts through word
+            if (word[i] != WORDEND) { // checks if word[i] is the end of a letter
                 temp[j] = word[i];
-            } else {
+            } else { // Once word is in temp searches BST for the letter the code is associated with
                 pNode node = letterSearch(temp, 0, root);
                 num[count] = node->letter;
                 count++;
@@ -49,27 +52,27 @@ void translate(char word[], int num[]) {
                 j = -1;
             }
         }
-    } else {
+    } else { // Assigns each letter to its decimal number
         for (int i = 0; i < ARR_SIZE && word[i + 1] != END; i++) {
-            num[i] = toupper(word[i]);
+            num[i] = (int)toupper(word[i]);
         }
     }
 }
 
-void toMorse(char code[], char word[], int received[]) {
-    for (int i = 0; i < ARR_SIZE && received[i] != END; i++) {
+void toMorse(char code[], char word[], int received[]) { // Translates deci array to morse code and text
+    for (int i = 0; i < ARR_SIZE && received[i] != END; i++) { // Traverses reviced array to assign to new array as a char 
         if (received[i] != END) {
             word[i] = (char)received[i];
         }
     }
     int pos = 0;
     path->top = -1;
-    for (int i = 0; i < ARR_SIZE && received[i] != END; i++) {
-        if (received[i] == SPACE)
+    for (int i = 0; i < ARR_SIZE && received[i] != END; i++) { // Traverses recived array to assing to new array in morse code
+        if (received[i] == SPACE)  // Check if the current value is a space
             code[pos++] = SPACE;
-        else if (codeSearch(root, received[i])) {
+        else if (codeSearch(root, received[i])) { // Search BST for the path to the letter of interest
             int inc = path->top;
-            for (int j = inc; j >= 0; j--) {
+            for (int j = inc; j >= 0; j--) { // Traverse stack to find path or morse code 
                 code[j + pos] = pop();
             }
             pos += inc + 1;
@@ -78,14 +81,15 @@ void toMorse(char code[], char word[], int received[]) {
     }
 }
 
-void push(int value) {
+void push(int value) { // Pushes values onto stack
     path->stack[++(path->top)] = value;
 }
 
-int pop() {
+int pop() { // Pops values from stack
     return path->stack[(path->top)--];
 }
 
+// Array of morse code definition
 int translation[26][6] = {{65, DOT, DASH, WORDEND, WORDEND, WORDEND},     // A
                           {66, DASH, DOT, DOT, DOT, WORDEND},             // B
                           {67, DASH, DOT, DASH, DOT, WORDEND},            // C
